# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atomic Patent API
  description: Machine-readable specification for the Atomic Patent Soroban smart contract interface.
  version: 1.0.0

tags:
  - name: IP Registry
    description: Commit and query intellectual property records
  - name: Atomic Swap
    description: Trustless patent sale via atomic swap

paths:
  /ip/commit:
    post:
      tags: [IP Registry]
      summary: Timestamp a new IP commitment
      operationId: commitIp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommitIpRequest'
      responses:
        '200':
          description: IP committed successfully, returns assigned ip_id
          content:
            application/json:
              schema:
                type: integer
                format: int64
        '400':
          description: Invalid request (zero hash, duplicate hash)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /ip/{ip_id}:
    get:
      tags: [IP Registry]
      summary: Retrieve an IP record by ID
      operationId: getIp
      parameters:
        - name: ip_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: IP record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IpRecord'
        '404':
          description: IP record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /ip/transfer:
    post:
      tags: [IP Registry]
      summary: Transfer IP ownership to a new address
      operationId: transferIp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferIpRequest'
      responses:
        '200':
          description: Ownership transferred successfully
        '404':
          description: IP record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /ip/verify:
    post:
      tags: [IP Registry]
      summary: Verify a Pedersen commitment
      operationId: verifyCommitment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyCommitmentRequest'
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyCommitmentResponse'
        '404':
          description: IP record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /ip/owner/{owner}:
    get:
      tags: [IP Registry]
      summary: List all IP IDs owned by a Stellar address
      operationId: listIpByOwner
      parameters:
        - name: owner
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of IP IDs (null if none)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListIpByOwnerResponse'

  /swap/initiate:
    post:
      tags: [Atomic Swap]
      summary: Seller initiates a patent sale
      operationId: initiateSwap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InitiateSwapRequest'
      responses:
        '200':
          description: Swap initiated, returns swap_id
          content:
            application/json:
              schema:
                type: integer
                format: int64
        '400':
          description: Seller is not IP owner or active swap exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /swap/{swap_id}/accept:
    post:
      tags: [Atomic Swap]
      summary: Buyer accepts a pending swap
      operationId: acceptSwap
      parameters:
        - name: swap_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptSwapRequest'
      responses:
        '200':
          description: Swap accepted
        '400':
          description: Swap not in Pending state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Swap not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /swap/{swap_id}/reveal:
    post:
      tags: [Atomic Swap]
      summary: Seller reveals the decryption key
      operationId: revealKey
      parameters:
        - name: swap_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevealKeyRequest'
      responses:
        '200':
          description: Key revealed, swap completed
        '400':
          description: Swap not in Accepted state or caller is not seller
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Swap not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /swap/{swap_id}/cancel:
    post:
      tags: [Atomic Swap]
      summary: Cancel a pending swap
      operationId: cancelSwap
      parameters:
        - name: swap_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelSwapRequest'
      responses:
        '200':
          description: Swap cancelled
        '400':
          description: Swap not in Pending state or canceller is not seller/buyer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Swap not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /swap/{swap_id}/cancel-expired:
    post:
      tags: [Atomic Swap]
      summary: Buyer cancels an Accepted swap after expiry
      operationId: cancelExpiredSwap
      parameters:
        - name: swap_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelExpiredSwapRequest'
      responses:
        '200':
          description: Expired swap cancelled
        '400':
          description: Swap not expired, not Accepted, or caller is not buyer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Swap not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /swap/{swap_id}:
    get:
      tags: [Atomic Swap]
      summary: Read a swap record by ID
      operationId: getSwap
      parameters:
        - name: swap_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Swap record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwapRecord'
        '404':
          description: Swap not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CommitIpRequest:
      type: object
      required: [owner, commitment_hash]
      properties:
        owner:
          type: string
          description: Stellar address of the IP owner (must sign the transaction)
        commitment_hash:
          type: string
          description: 32-byte Pedersen commitment hash, hex-encoded

    IpRecord:
      type: object
      required: [ip_id, owner, commitment_hash, timestamp, revoked]
      properties:
        ip_id:
          type: integer
          format: int64
        owner:
          type: string
        commitment_hash:
          type: string
        timestamp:
          type: integer
          format: int64
        revoked:
          type: boolean
          description: Whether the IP record has been revoked

    TransferIpRequest:
      type: object
      required: [ip_id, new_owner]
      properties:
        ip_id:
          type: integer
          format: int64
        new_owner:
          type: string

    VerifyCommitmentRequest:
      type: object
      required: [ip_id, secret, blinding_factor]
      properties:
        ip_id:
          type: integer
          format: int64
        secret:
          type: string
          description: 32-byte secret, hex-encoded
        blinding_factor:
          type: string
          description: 32-byte blinding factor, hex-encoded

    VerifyCommitmentResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
          description: true if sha256(secret || blinding_factor) matches the stored commitment hash

    ListIpByOwnerResponse:
      type: object
      properties:
        ip_ids:
          type: array
          items:
            type: integer
            format: int64
          nullable: true

    SwapStatus:
      type: string
      enum: [Pending, Accepted, Completed, Cancelled]

    SwapRecord:
      type: object
      required: [ip_id, ip_registry_id, seller, buyer, price, token, status, expiry]
      properties:
        ip_id:
          type: integer
          format: int64
        ip_registry_id:
          type: string
        seller:
          type: string
        buyer:
          type: string
        price:
          type: integer
          format: int64
          description: Price in stroops (1 XLM = 10_000_000 stroops)
        token:
          type: string
        status:
          $ref: '#/components/schemas/SwapStatus'
        expiry:
          type: integer
          format: int64
          description: Ledger timestamp after which buyer may cancel an Accepted swap

    InitiateSwapRequest:
      type: object
      required: [ip_registry_id, ip_id, seller, price, buyer, token]
      properties:
        ip_registry_id:
          type: string
        ip_id:
          type: integer
          format: int64
        seller:
          type: string
        price:
          type: integer
          format: int64
        buyer:
          type: string
        token:
          type: string
          description: Stellar asset contract address for the payment token

    AcceptSwapRequest:
      type: object
      required: [buyer]
      properties:
        buyer:
          type: string

    RevealKeyRequest:
      type: object
      required: [caller, secret, blinding_factor]
      properties:
        caller:
          type: string
        secret:
          type: string
          description: 32-byte secret, hex-encoded
        blinding_factor:
          type: string
          description: 32-byte blinding factor, hex-encoded

    CancelSwapRequest:
      type: object
      required: [canceller]
      properties:
        canceller:
          type: string

    CancelExpiredSwapRequest:
      type: object
      required: [caller]
      properties:
        caller:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string